Create Car Loan Payment Calculator - Project
This is a simple Car Loan Payment Calculator without a gui connected.
Simply fill out the integer variables to what your car loan is, how
long it lasts, the interest rate, and how much of a down payment
you have put down. Based on that information it will figure
out how much you need to pay on a monthly basis.